#jira UE-75803 Fix bad signed/unsigned logic causing crash when adjusting RVT mip level removal #lockdown Marcus.Wassmer #rb none #rnx
[CL 6911372 by Jeremy Moore in Dev-Rendering branch]

+++ Engine/Source/Runtime/Engine/Private/VT/RuntimeVirtualTexture.cpp
@@ -183,7 +183,7 @@ void URuntimeVirtualTexture::GetProducerDescription(FVTProducerDescription& OutD
 	OutDesc.TileBorderSize = GetTileBorderSize();
 	OutDesc.BlockWidthInTiles = GetWidth() / GetTileSize();
 	OutDesc.BlockHeightInTiles = GetHeight() / GetTileSize();
-	OutDesc.MaxLevel = FMath::Max(FMath::CeilLogTwo(FMath::Max(OutDesc.BlockWidthInTiles, OutDesc.BlockHeightInTiles)) - RemoveLowMips, 1u);
+	OutDesc.MaxLevel = FMath::Max((int32)FMath::CeilLogTwo(FMath::Max(OutDesc.BlockWidthInTiles, OutDesc.BlockHeightInTiles)) - RemoveLowMips, 1);
 	OutDesc.DepthInTiles = 1;
 	OutDesc.WidthInBlocks = 1;
 	OutDesc.HeightInBlocks = 1;
